Split Qwen35 alternate runtime dispatch
